+++ src/main/java/nom/bdezonia/zorbage/type/real/float16/Float16CartesianTensorProductMember.java
@@ -702,7 +702,7 @@ public final class Float16CartesianTensorProductMember
 		if (component == 0) {
 			Float16Member tmp = tmpFloat.get();
 			getV(index, tmp);
-			return BigDecimal.valueOf(tmp.v()).toBigInteger();
+			return BigInteger.valueOf((long) tmp.v());
 		}
 		return BigInteger.ZERO;
 	}
@@ -800,7 +800,7 @@ public final class Float16CartesianTensorProductMember
 		else {
 			Float16Member tmp = tmpFloat.get();
 			getV(index, tmp);
-			return BigDecimal.valueOf(tmp.v()).toBigInteger();
+			return BigInteger.valueOf((long) tmp.v());
 		}
 	}
 
@@ -1074,7 +1074,7 @@ public final class Float16CartesianTensorProductMember
 		BigInteger[] values = new BigInteger[1 * (int) storage.size()];
 		for (int i = 0; i < storage.size(); i++) {
 			storage.get(i, value);
-			values[i] = BigDecimal.valueOf(value.v()).toBigInteger();
+			values[i] = BigInteger.valueOf((long) value.v());
 		}
 		return values;
 	}
